+++ bridges/bin/runtime-common/src/messages_benchmarking.rs
@@ -19,25 +19,42 @@
 
 #![cfg(feature = "runtime-benchmarks")]
 
-use crate::{
-	messages::{AccountIdOf, BridgedChain, HashOf, MessageBridge, ThisChain},
-	messages_generation::{
-		encode_all_messages, encode_lane_data, prepare_message_delivery_storage_proof,
-		prepare_messages_storage_proof,
-	},
-};
+// use crate::{
+// 	messages::{BridgedChain, HashOf, ThisChain},
+// 	messages_generation::{
+// 		encode_all_messages, encode_lane_data, prepare_message_delivery_storage_proof,
+// 		prepare_messages_storage_proof,
+// 	},
+// };
 
 use bp_messages::{
 	source_chain::FromBridgedChainMessagesDeliveryProof,
 	target_chain::FromBridgedChainMessagesProof, MessagePayload,
 };
 use bp_polkadot_core::parachains::ParaHash;
-use bp_runtime::{Chain, Parachain, StorageProofSize, UnderlyingChainOf};
+use bp_runtime::{Chain, Parachain, StorageProofSize};
+// TODO: https://github.com/paritytech/parity-bridges-common/issues/1666 - merge with message
+// generation methods from messages pallet and use it here
+
+use bp_messages::{storage_keys, ChainWithMessages};
+use bp_runtime::{
+	grow_trie_leaf_value, record_all_trie_keys, AccountIdOf, HashOf, HasherOf, UntrustedVecDb,
+};
 use codec::Encode;
 use frame_support::weights::Weight;
-use pallet_bridge_messages::benchmarking::{MessageDeliveryProofParams, MessageProofParams};
-use sp_runtime::traits::{Header, Zero};
+use pallet_bridge_messages::{
+	benchmarking::{MessageDeliveryProofParams, MessageProofParams},
+	messages_generation::{encode_all_messages, encode_lane_data, prepare_messages_storage_proof},
+	BridgedChainOf, ThisChainOf,
+};
+use sp_runtime::{
+	traits::{Header, Zero},
+	StateVersion,
+};
 use sp_std::prelude::*;
+use sp_trie::{
+	LayoutV0, LayoutV1, MemoryDB, StorageProof, TrieConfiguration, TrieDBMutBuilder, TrieMut,
+};
 use xcm::latest::prelude::*;
 
 /// Prepare inbound bridge message according to given message proof parameters.
@@ -75,27 +92,32 @@ fn prepare_inbound_message(
 /// This method is intended to be used when benchmarking pallet, linked to the chain that
 /// uses GRANDPA finality. For parachains, please use the `prepare_message_proof_from_parachain`
 /// function.
-pub fn prepare_message_proof_from_grandpa_chain<R, FI, B>(
+pub fn prepare_message_proof_from_grandpa_chain<R, FI, MI>(
 	params: MessageProofParams,
 	message_generator: impl Fn(usize) -> MessagePayload,
-) -> (FromBridgedChainMessagesProof<HashOf<BridgedChain<B>>>, Weight)
+) -> (FromBridgedChainMessagesProof<HashOf<BridgedChainOf<R, MI>>>, Weight)
 where
-	R: pallet_bridge_grandpa::Config<FI, BridgedChain = UnderlyingChainOf<BridgedChain<B>>>,
+	R: pallet_bridge_grandpa::Config<FI, BridgedChain = BridgedChainOf<R, MI>>
+		+ pallet_bridge_messages::Config<
+			MI,
+			BridgedHeaderChain = pallet_bridge_grandpa::Pallet<R, FI>,
+		>,
 	FI: 'static,
-	B: MessageBridge,
+	MI: 'static,
 {
 	// prepare storage proof
-	let (state_root, storage) = prepare_messages_storage_proof::<B>(
-		params.lane,
-		params.message_nonces.clone(),
-		params.outbound_lane_data.clone(),
-		params.size,
-		prepare_inbound_message(&params, message_generator),
-		encode_all_messages,
-		encode_lane_data,
-		false,
-		false,
-	);
+	let (state_root, storage) =
+		prepare_messages_storage_proof::<BridgedChainOf<R, MI>, ThisChainOf<R, MI>>(
+			params.lane,
+			params.message_nonces.clone(),
+			params.outbound_lane_data.clone(),
+			params.size,
+			|_| prepare_inbound_message(&params, &message_generator),
+			encode_all_messages,
+			encode_lane_data,
+			false,
+			false,
+		);
 
 	// update runtime storage
 	let (_, bridged_header_hash) = insert_header_to_grandpa_pallet::<R, FI>(state_root);
@@ -120,32 +142,33 @@ where
 /// This method is intended to be used when benchmarking pallet, linked to the chain that
 /// uses parachain finality. For GRANDPA chains, please use the
 /// `prepare_message_proof_from_grandpa_chain` function.
-pub fn prepare_message_proof_from_parachain<R, PI, B>(
+pub fn prepare_message_proof_from_parachain<R, PI, MI>(
 	params: MessageProofParams,
 	message_generator: impl Fn(usize) -> MessagePayload,
-) -> (FromBridgedChainMessagesProof<HashOf<BridgedChain<B>>>, Weight)
+) -> (FromBridgedChainMessagesProof<HashOf<BridgedChainOf<R, MI>>>, Weight)
 where
-	R: pallet_bridge_parachains::Config<PI>,
+	R: pallet_bridge_parachains::Config<PI> + pallet_bridge_messages::Config<MI>,
 	PI: 'static,
-	B: MessageBridge,
-	UnderlyingChainOf<BridgedChain<B>>: Chain<Hash = ParaHash> + Parachain,
+	MI: 'static,
+	BridgedChainOf<R, MI>: Chain<Hash = ParaHash> + Parachain,
 {
 	// prepare storage proof
-	let (state_root, storage) = prepare_messages_storage_proof::<B>(
-		params.lane,
-		params.message_nonces.clone(),
-		params.outbound_lane_data.clone(),
-		params.size,
-		prepare_inbound_message(&params, message_generator),
-		encode_all_messages,
-		encode_lane_data,
-		false,
-		false,
-	);
+	let (state_root, storage) =
+		prepare_messages_storage_proof::<BridgedChainOf<R, MI>, ThisChainOf<R, MI>>(
+			params.lane,
+			params.message_nonces.clone(),
+			params.outbound_lane_data.clone(),
+			params.size,
+			|_| prepare_inbound_message(&params, &message_generator),
+			encode_all_messages,
+			encode_lane_data,
+			false,
+			false,
+		);
 
 	// update runtime storage
 	let (_, bridged_header_hash) =
-		insert_header_to_parachains_pallet::<R, PI, UnderlyingChainOf<BridgedChain<B>>>(state_root);
+		insert_header_to_parachains_pallet::<R, PI, BridgedChainOf<R, MI>>(state_root);
 
 	(
 		FromBridgedChainMessagesProof {
@@ -164,21 +187,21 @@ where
 /// This method is intended to be used when benchmarking pallet, linked to the chain that
 /// uses GRANDPA finality. For parachains, please use the
 /// `prepare_message_delivery_proof_from_parachain` function.
-pub fn prepare_message_delivery_proof_from_grandpa_chain<R, FI, B>(
-	params: MessageDeliveryProofParams<AccountIdOf<ThisChain<B>>>,
-) -> FromBridgedChainMessagesDeliveryProof<HashOf<BridgedChain<B>>>
+pub fn prepare_message_delivery_proof_from_grandpa_chain<R, FI, MI>(
+	params: MessageDeliveryProofParams<AccountIdOf<ThisChainOf<R, MI>>>,
+) -> FromBridgedChainMessagesDeliveryProof<HashOf<BridgedChainOf<R, MI>>>
 where
-	R: pallet_bridge_grandpa::Config<FI, BridgedChain = UnderlyingChainOf<BridgedChain<B>>>,
+	R: pallet_bridge_grandpa::Config<FI, BridgedChain = BridgedChainOf<R, MI>>
+		+ pallet_bridge_messages::Config<
+			MI,
+			BridgedHeaderChain = pallet_bridge_grandpa::Pallet<R, FI>,
+		>,
 	FI: 'static,
-	B: MessageBridge,
+	MI: 'static,
 {
 	// prepare storage proof
 	let lane = params.lane;
-	let (state_root, storage_proof) = prepare_message_delivery_storage_proof::<B>(
-		params.lane,
-		params.inbound_lane_data,
-		params.size,
-	);
+	let (state_root, storage_proof) = prepare_message_delivery_proof::<R, MI>(params);
 
 	// update runtime storage
 	let (_, bridged_header_hash) = insert_header_to_grandpa_pallet::<R, FI>(state_root);
@@ -195,32 +218,91 @@ where
 /// This method is intended to be used when benchmarking pallet, linked to the chain that
 /// uses parachain finality. For GRANDPA chains, please use the
 /// `prepare_message_delivery_proof_from_grandpa_chain` function.
-pub fn prepare_message_delivery_proof_from_parachain<R, PI, B>(
-	params: MessageDeliveryProofParams<AccountIdOf<ThisChain<B>>>,
-) -> FromBridgedChainMessagesDeliveryProof<HashOf<BridgedChain<B>>>
+pub fn prepare_message_delivery_proof_from_parachain<R, PI, MI>(
+	params: MessageDeliveryProofParams<AccountIdOf<ThisChainOf<R, MI>>>,
+) -> FromBridgedChainMessagesDeliveryProof<HashOf<BridgedChainOf<R, MI>>>
 where
-	R: pallet_bridge_parachains::Config<PI>,
+	R: pallet_bridge_parachains::Config<PI> + pallet_bridge_messages::Config<MI>,
 	PI: 'static,
-	B: MessageBridge,
-	UnderlyingChainOf<BridgedChain<B>>: Chain<Hash = ParaHash> + Parachain,
+	MI: 'static,
+	BridgedChainOf<R, MI>: Chain<Hash = ParaHash> + Parachain,
 {
 	// prepare storage proof
 	let lane = params.lane;
-	let (state_root, storage_proof) = prepare_message_delivery_storage_proof::<B>(
-		params.lane,
-		params.inbound_lane_data,
-		params.size,
-	);
+	let (state_root, storage_proof) = prepare_message_delivery_proof::<R, MI>(params);
 
 	// update runtime storage
 	let (_, bridged_header_hash) =
-		insert_header_to_parachains_pallet::<R, PI, UnderlyingChainOf<BridgedChain<B>>>(state_root);
+		insert_header_to_parachains_pallet::<R, PI, BridgedChainOf<R, MI>>(state_root);
 
 	FromBridgedChainMessagesDeliveryProof {
 		bridged_header_hash: bridged_header_hash.into(),
 		storage_proof,
 		lane,
 	}
+}
+
+/// Prepare in-memory message delivery proof, without inserting anything to the runtime storage.
+fn prepare_message_delivery_proof<R, MI>(
+	params: MessageDeliveryProofParams<AccountIdOf<ThisChainOf<R, MI>>>,
+) -> (HashOf<BridgedChainOf<R, MI>>, UntrustedVecDb)
+where
+	R: pallet_bridge_messages::Config<MI>,
+	MI: 'static,
+{
+	match BridgedChainOf::<R, MI>::STATE_VERSION {
+		StateVersion::V0 =>
+			do_prepare_message_delivery_proof::<R, MI, LayoutV0<HasherOf<BridgedChainOf<R, MI>>>>(
+				params,
+			),
+		StateVersion::V1 =>
+			do_prepare_message_delivery_proof::<R, MI, LayoutV1<HasherOf<BridgedChainOf<R, MI>>>>(
+				params,
+			),
+	}
+}
+
+/// Prepare in-memory message delivery proof, without inserting anything to the runtime storage.
+fn do_prepare_message_delivery_proof<
+	R,
+	MI,
+	L: TrieConfiguration<Hash = HasherOf<BridgedChainOf<R, MI>>>,
+>(
+	params: MessageDeliveryProofParams<AccountIdOf<ThisChainOf<R, MI>>>,
+) -> (HashOf<BridgedChainOf<R, MI>>, UntrustedVecDb)
+where
+	R: pallet_bridge_messages::Config<MI>,
+	MI: 'static,
+{
+	// prepare Bridged chain storage with inbound lane state
+	let storage_key = storage_keys::inbound_lane_data_key(
+		R::ThisChain::WITH_CHAIN_MESSAGES_PALLET_NAME,
+		&params.lane,
+	)
+	.0;
+	let mut root = Default::default();
+	let mut mdb = MemoryDB::default();
+	{
+		let mut trie = TrieDBMutBuilder::<L>::new(&mut mdb, &mut root).build();
+		let inbound_lane_data =
+			grow_trie_leaf_value(params.inbound_lane_data.encode(), params.size);
+		trie.insert(&storage_key, &inbound_lane_data)
+			.map_err(|_| "TrieMut::insert has failed")
+			.expect("TrieMut::insert should not fail in benchmarks");
+	}
+
+	// generate storage proof to be delivered to This chain
+	let read_proof = record_all_trie_keys::<L, _>(&mdb, &root)
+		.map_err(|_| "record_all_trie_keys has failed")
+		.expect("record_all_trie_keys should not fail in benchmarks");
+	let storage_proof = UntrustedVecDb::try_new::<HasherOf<BridgedChainOf<R, MI>>>(
+		StorageProof::new(read_proof),
+		root,
+		vec![storage_key],
+	)
+	.unwrap();
+
+	(root, storage_proof)
 }
 
 /// Insert header to the bridge GRANDPA pallet.
