+++ bridges/modules/messages/Cargo.toml
@@ -18,6 +18,7 @@ scale-info = { version = "2.11.1", default-features = false, features = ["derive
 
 # Bridge dependencies
 
+bp-header-chain = { path = "../../primitives/header-chain", default-features = false }
 bp-messages = { path = "../../primitives/messages", default-features = false }
 bp-runtime = { path = "../../primitives/runtime", default-features = false }
 
@@ -28,15 +29,20 @@ frame-support = { path = "../../../substrate/frame/support", default-features = 
 frame-system = { path = "../../../substrate/frame/system", default-features = false }
 sp-runtime = { path = "../../../substrate/primitives/runtime", default-features = false }
 sp-std = { path = "../../../substrate/primitives/std", default-features = false }
+sp-trie = { path = "../../../substrate/primitives/trie", default-features = false, optional = true }
 
 [dev-dependencies]
 bp-test-utils = { path = "../../primitives/test-utils" }
+pallet-bridge-grandpa = { path = "../grandpa" }
 pallet-balances = { path = "../../../substrate/frame/balances" }
 sp-io = { path = "../../../substrate/primitives/io" }
+sp-core = { path = "../../../substrate/primitives/core" }
+sp-trie = { path = "../../../substrate/primitives/trie" }
 
 [features]
 default = ["std"]
 std = [
+	"bp-header-chain/std",
 	"bp-messages/std",
 	"bp-runtime/std",
 	"codec/std",
@@ -48,6 +54,7 @@ std = [
 	"scale-info/std",
 	"sp-runtime/std",
 	"sp-std/std",
+	"sp-trie/std"
 ]
 runtime-benchmarks = [
 	"frame-benchmarking/runtime-benchmarks",
@@ -61,4 +68,7 @@ try-runtime = [
 	"frame-system/try-runtime",
 	"pallet-balances/try-runtime",
 	"sp-runtime/try-runtime",
+]
+test-helpers = [
+	"sp-trie",
 ]
