+++ bridges/primitives/runtime/src/lib.rs
@@ -41,7 +41,7 @@ pub use chain::{
 pub use frame_support::storage::storage_prefix as storage_value_final_key;
 use num_traits::{CheckedAdd, CheckedSub, One, SaturatingAdd, Zero};
 pub use storage_proof::{
-	record_all_keys as record_all_trie_keys, Error as StorageProofError,
+	grow_trie_leaf_value, record_all_keys as record_all_trie_keys, Error as StorageProofError,
 	ProofSize as StorageProofSize, RawStorageProof, StorageProofChecker,
 };
 pub use storage_types::BoundedStorageValue;
